Add support for CC5X and PK2
- Add new platform class PlatformPk2Cc5x
- Add platform to programwindow
- Add syntax highlighting

This adds support for a "tool chain" built around B. Knudsen's CC5X
compiler for midrange PIC devices. It is used together with PK2CMD from
Microchip to download the program to the device.

Since two separate programs are required, the support is built around
simple scripts (bat/sh) that will execute the programs in order. Those
scripts and instructions on how to use them are available under the
download link.